src/test/javascript/spec/app/entities/song/song-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';

import * as config from '@/shared/config/config';
import SongUpdateComponent from '@/entities/song/song-update.vue';
import SongClass from '@/entities/song/song-update.component';
import SongService from '@/entities/song/song.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Song Management Update Component', () => {
    let wrapper: Wrapper<SongClass>;
    let comp: SongClass;
    let songServiceStub: SinonStubbedInstance<SongService>;

    beforeEach(() => {
      songServiceStub = sinon.createStubInstance<SongService>(SongService);

      wrapper = shallowMount<SongClass>(SongUpdateComponent, {
        store,
        localVue,
        router,
        provide: {
          songService: () => songServiceStub,
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.song = entity;
        songServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(songServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.song = entity;
        songServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(songServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundSong = { id: 123 };
        songServiceStub.find.resolves(foundSong);
        songServiceStub.retrieve.resolves([foundSong]);

        // WHEN
        comp.beforeRouteEnter({ params: { songId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.song).toBe(foundSong);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
